// src/generated/api/checkout/models/listAdminShippingConsultationsStatus.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */

export type ListAdminShippingConsultationsStatus =
  (typeof ListAdminShippingConsultationsStatus)[keyof typeof ListAdminShippingConsultationsStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ListAdminShippingConsultationsStatus = {
  AWAITING_SHIPPING_CONSULTATION: 'AWAITING_SHIPPING_CONSULTATION',
  QUOTED: 'QUOTED',
} as const;